Some reconfig for get_mem_block

+++ paxforth/src/runner/wasm/wasmtime.rs
@@ -7,7 +7,7 @@ use wasmtime::*;
 pub struct WasmContext {
     instance: Instance,
     buffer: Rc<Mutex<Vec<u8>>>,
-    memory: Rc<Mutex<HashMap<u32, u16>>>,
+    memory: Rc<Mutex<HashMap<usize, u16>>>,
 }
 
 impl WasmContext {
@@ -30,14 +30,18 @@ impl WasmContext {
         Ok(buffer_ref.iter().cloned().collect::<Vec<u8>>())
     }
 
+    /**
+     * Memory modification functions.
+     */
+
     pub fn get_mem(&self, index: u32) -> anyhow::Result<u16> {
         let memory_ref = self.memory.lock().unwrap();
-        Ok(memory_ref.get(&index).map(|x| *x).unwrap_or(0))
+        Ok(memory_ref.get(&(index as usize)).map(|x| *x).unwrap_or(0))
     }
 
     pub fn set_mem(&self, index: u32, value: u16) -> anyhow::Result<()> {
         let mut memory_ref = self.memory.lock().unwrap();
-        memory_ref.insert(index, value);
+        memory_ref.insert(index as usize, value);
         Ok(())
     }
 }
@@ -48,7 +52,7 @@ pub fn parse_wasm(binary: &[u8], export_buffer: bool) -> anyhow::Result<WasmCont
 
     // Imports.
     let buffer: Rc<Mutex<Vec<u8>>> = Rc::new(Mutex::new(Vec::new()));
-    let memory: Rc<Mutex<HashMap<u32, u16>>> = Rc::new(Mutex::new(HashMap::new()));
+    let memory: Rc<Mutex<HashMap<usize, u16>>> = Rc::new(Mutex::new(HashMap::new()));
     let print_func = Func::wrap(&store, {
         let buffer = buffer.clone();
         move |value: i32| -> i32 {
@@ -74,14 +78,22 @@ pub fn parse_wasm(binary: &[u8], export_buffer: bool) -> anyhow::Result<WasmCont
         let memory = memory.clone();
         move |address: i32| -> i32 {
             let address = (address as u32) & 0xFFFF;
-            *memory.lock().unwrap().get(&address).unwrap_or(&0) as _
+            *memory
+                .lock()
+                .unwrap()
+                .get(&(address as usize))
+                .unwrap_or(&0) as _
         }
     });
     let extmem_load_8_func = Func::wrap(&store, {
         let memory = memory.clone();
         move |address: i32| -> i32 {
             let address = (address as u32) & 0xFFFF;
-            *memory.lock().unwrap().get(&address).unwrap_or(&0) as u8 as _
+            *memory
+                .lock()
+                .unwrap()
+                .get(&(address as usize))
+                .unwrap_or(&0) as u8 as _
         }
     });
     let extmem_store_func = Func::wrap(&store, {
@@ -89,7 +101,10 @@ pub fn parse_wasm(binary: &[u8], export_buffer: bool) -> anyhow::Result<WasmCont
         move |address: i32, value: i32| -> () {
             let address = (address as u32) & 0xFFFF;
             let value = (value as u32) & 0xFFFF;
-            memory.lock().unwrap().insert(address, value as u16);
+            memory
+                .lock()
+                .unwrap()
+                .insert((address as usize), value as u16);
         }
     });
     let extmem_store_8_func = Func::wrap(&store, {
@@ -97,7 +112,10 @@ pub fn parse_wasm(binary: &[u8], export_buffer: bool) -> anyhow::Result<WasmCont
         move |address: i32, value: i32| -> () {
             let address = (address as u32) & 0xFFFF;
             let value = (value as u32) & 0xFFFF;
-            memory.lock().unwrap().insert(address, value as u8 as u16);
+            memory
+                .lock()
+                .unwrap()
+                .insert((address as usize), value as u8 as u16);
         }
     });
